The regenerating code must be working now.
Fixed a few problems with the regenerating code decoder in ReedSolomonCode.java
Earlier, if there were muliple erasures, a reed solomon decoder was direclty
initiated but now it first checks whether those erasures can be fixed just by
XORing. One thing we may need to check is for making sure that the input parameters
for stripeSize, paritySize and simpleParityDegree are valid. Also at this point,
it is difficult to check whether the code performs pure RS decoding if the
paritySizeSRC is 0. This is because we are using simpleParityDegree instead of
paritySizeSRC as an input argument. Could possibly change it.