Removed extra parity block
The RS parities are such that the sum of all the RS parities and the
data blocks add to zero, which means that the sum of all the simple
parities also add to zero. So it is possible to remove one extra parity
and still be able to perform the same as before. Here, we have just
removed the first parity. In hdfs-site.xml, input the paritySizeSRC
as one less than the number of SRC's required. If paritySizeSRC is
already zero, leave it at zero.